Dedup mocking code
